% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{AbfAvailEpisodeIdx}
\alias{AbfAvailEpisodeIdx}
\title{AbfAvailEpisodeIdx same as AvailEpisodeIdx but operates on an abf data}
\usage{
AbfAvailEpisodeIdx(abf)
}
\arguments{
\item{abf}{data loaded from abf file}
}
\value{
a vector of available indices
}
\description{
AbfAvailEpisodeIdx same as AvailEpisodeIdx but operates on an abf data
}
\examples{
idx <- AbfAvailEpisodeIdx(abf)
}
