% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{AbfAvailEpisodeName}
\alias{AbfAvailEpisodeName}
\title{AbfAvailEpisodeName same as AvailEpisodeName but operates on an abf data}
\usage{
AbfAvailEpisodeName(abf)
}
\arguments{
\item{abf}{data loaded from abf file}
}
\value{
a vector of available names
}
\description{
AbfAvailEpisodeName same as AvailEpisodeName but operates on an abf data
}
\examples{
epi_names <- AbfAvailEpisodeName(abf)
}
