% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{AbfRemoveEpisode}
\alias{AbfRemoveEpisode}
\title{AbfRemoveEpisode removes episodes from all channels of abf}
\usage{
AbfRemoveEpisode(abf, episode_idx)
}
\arguments{
\item{abf}{data loaded from abf file}

\item{episode_idx}{a list or vector of episode indices to be removed}
}
\value{
abf data with desired episodes removed
}
\description{
AbfRemoveEpisode removes episodes from all channels of abf
}
\examples{
abf <- AbfRemoveEpisode(abf, c(1,2,3,4,8))
}
