% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{AllSamples_ChannelMeans}
\alias{AllSamples_ChannelMeans}
\title{AllSamples_ChannelMeans calculates average channel data for every episode within the given interval}
\usage{
AllSamples_ChannelMeans(abf_list, intv_list, chan_id)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{intv_list}{a list of intervals}

\item{chan_id}{channel id to calculate}
}
\value{
a data.frame of the calculated means.
}
\description{
AllSamples_ChannelMeans calculates average channel data for every episode within the given interval
}
