% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{AllSamples_CurrentMeans}
\alias{AllSamples_CurrentMeans}
\title{AllSamples_VoltageMeans calculates average current for every episode within the given interval}
\usage{
AllSamples_CurrentMeans(abf_list, intv_list)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{intv_list}{a list of intervals}
}
\value{
a data.frame of the calculated means.
}
\description{
AllSamples_VoltageMeans calculates average current for every episode within the given interval
}
