% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{AllSamples_IRaw}
\alias{AllSamples_IRaw}
\title{AllSamples_IRaw aggregates all abf CURRENT data into a list of data.frame}
\usage{
AllSamples_IRaw(abf_list, intv_list)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{intv_list}{a list of corresponding sampling intervals}
}
\value{
A list of data.frame, each element correspond to an abf CURRENT data.
}
\description{
AllSamples_IRaw aggregates all abf CURRENT data into a list of data.frame
}
