% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{AllSamples_IVRaw}
\alias{AllSamples_IVRaw}
\title{AllSamples_IVRaw aggregates all abf data into a list of data.frame}
\usage{
AllSamples_IVRaw(abf_list, intv_list)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{intv_list}{a list of corresponding sampling intervals}
}
\value{
A list of data.frame, each element correspond to an abf data.
}
\description{
AllSamples_IVRaw aggregates all abf data into a list of data.frame
}
