% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{AllSamples_IVSummary}
\alias{AllSamples_IVSummary}
\title{AllSamples_IVSummary calculates average current and voltage for all samples listed in abf_list}
\usage{
AllSamples_IVSummary(abf_list, intv_list)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{intv_list}{a list of intervals}
}
\value{
a data.frame with Voltage, SEM Voltage, Current, SEM Current values
}
\description{
AllSamples_IVSummary calculates average current and voltage for all samples listed in abf_list
}
