% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{ChangeInterval}
\alias{ChangeInterval}
\title{ChangeInterval changes interval list manually}
\usage{
ChangeInterval(intv_list, sample_id, intv_start, intv_length)
}
\arguments{
\item{intv_list}{interval list}

\item{sample_id}{sample id}

\item{intv_start}{interval start position}

\item{intv_length}{interval length}
}
\value{
a list of intervals that is changed
}
\description{
ChangeInterval changes interval list manually
}
\examples{
intv_list <- ChangeInterval(intv_list, 1, 6000, 200)
}
