% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{episode.avail}
\alias{episode.avail}
\title{Get available episodes from given channel}
\usage{
episode.avail(df)
}
\arguments{
\item{df}{a channel}
}
\value{
availavle episode indices as vector of integer
}
\description{
Get available episodes from given channel
}
