% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{episode.get_idx}
\alias{episode.get_idx}
\title{Get episode indices from episode names}
\usage{
episode.get_idx(episode_names)
}
\arguments{
\item{episode_names}{}
}
\value{
episode indices as vector of numeric
}
\description{
Get episode indices from episode names
}
\examples{
episode.get_idx(c("epi5", "epi4"))
}
