% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{episode.get_names}
\alias{episode.get_names}
\title{Get episode names from episode indices}
\usage{
episode.get_names(episode_idx)
}
\arguments{
\item{episode_idx}{}
}
\value{
episode names as vector of characters
}
\description{
Get episode names from episode indices
}
\examples{
episode.get_names(c(1,3,5,8))
}
