% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{episode.remove_multi}
\alias{episode.remove_multi}
\title{Remove episodes from multiple channels, returns a list of data.frame same size of channels}
\usage{
episode.remove_multi(channels, episodes_to_remove)
}
\arguments{
\item{channels}{a list of channels}

\item{episodes_to_remove}{a vector of episode indices}
}
\value{
a list of channels with the desired episodes removed
}
\description{
Remove episodes from multiple channels, returns a list of data.frame same size of channels
}
