% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{episode.remove_single}
\alias{episode.remove_single}
\title{Remove episodes from single channel, returns a data.frame}
\usage{
episode.remove_single(channel, episodes_to_remove)
}
\arguments{
\item{channel}{a channel}

\item{episodes_to_remove}{a vector of episode indices}
}
\value{
a channel with the desired episodes removed
}
\description{
Remove episodes from single channel, returns a data.frame
}
