% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{EpisodeIdx}
\alias{EpisodeIdx}
\title{EpisodeIdx returns episode indices from a list or vector of episode names}
\usage{
EpisodeIdx(episode_names)
}
\arguments{
\item{episode_names}{a list or vector of episode names}
}
\value{
a list or vector of episode names, the return type is determined by typeof episode_names
}
\description{
EpisodeIdx returns episode indices from a list or vector of episode names
}
\examples{
epi_idx <- EpisodeIdx(c("epi3", "epi5", "epi6"))
}
