% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{EpisodeName}
\alias{EpisodeName}
\title{EpisodeName returns episode names from a list or vector of episode indices}
\usage{
EpisodeName(episode_idx)
}
\arguments{
\item{episode_idx}{a list or vector of episode indices}
}
\value{
a list or vector of episode names, the return type is determined by typeof episode_idx
}
\description{
EpisodeName returns episode names from a list or vector of episode indices
}
\examples{
epi_name <- EpisodeName(c(1,3,4,5))
}
