% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{FindAllSamplingInterval}
\alias{FindAllSamplingInterval}
\title{FindAllSampleInterval deos FindSamplingInterval for a list of abf data}
\usage{
FindAllSamplingInterval(abf_list, current_chan_id = 0, voltage_chan_id = 0,
  interval_size = 0, max_interval_expansion_rate = 1,
  allowed_voltage_delta = 0, epoch_name = "auto")
}
\arguments{
\item{abf_list}{a list of abf data}

\item{current_chan_id}{OPTIONAL, current channel id, usually can be identified automatically}

\item{voltage_chan_id}{OPTIONAL, voltage channel id, usually can be identified automatically}

\item{interval_size}{OPTIONAL, prefered minimum interval size.}

\item{max_interval_expansion_rate}{OPTIONAL, determine if the function returns variable interval size.}

\item{allowed_voltage_delta}{OPTIONAL, allowed maximum deviation of voltage W.R.T protocol setting in the sampling interval.}

\item{epoch_name}{OPTIONAL, determines which in which epoch to find the interval.}
}
\value{
a list of intervals found
}
\description{
FindAllSampleInterval deos FindSamplingInterval for a list of abf data
}
