% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{gapfree.current_avg}
\alias{gapfree.current_avg}
\title{Calculate average current from a gapfree abfdata by intervals}
\usage{
gapfree.current_avg(abfdata, n, time_unit)
}
\arguments{
\item{n}{length of each interval}

\item{time_unit}{time unit used, can be "ms", "s" or "ticks"}
}
\description{
Calculate average current from a gapfree abfdata by intervals
}
