% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot_Channel}
\alias{Plot_Channel}
\title{Plot_Channel plots a single channel}
\usage{
Plot_Channel(abf, chan_id)
}
\arguments{
\item{abf}{data loaded from abf file}

\item{chan_id}{channel id to plot}
}
\value{
a ggplot object
}
\description{
Plot_Channel plots a single channel
}
\examples{
Plot_Channel(abf, 2)
}
