% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot_ChannelWithIntv}
\alias{Plot_ChannelWithIntv}
\title{Plot_ChannelWithIntv plots a single channel with two dotted lines indicating an interval}
\usage{
Plot_ChannelWithIntv(abf, chan_id, intv)
}
\arguments{
\item{abf}{data loaded from abf file}

\item{chan_id}{channel id to plot}

\item{intv}{a vector of 2 numerics describing an interval}
}
\value{
a ggplot object
}
\description{
Plot_ChannelWithIntv plots a single channel with two dotted lines indicating an interval
}
\examples{
intv <- FindSamplingInterval(abf); Plot_ChannelWithIntv(abf, 1, intv)
}
