% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot_IVChannel}
\alias{Plot_IVChannel}
\title{Plot_IVChannel plots both current and voltage channels on a same graph}
\usage{
Plot_IVChannel(abf)
}
\arguments{
\item{abf}{data loaded from abf file}
}
\value{
a ggplot object
}
\description{
Plot_IVChannel plots both current and voltage channels on a same graph
}
\examples{
Plot_IVChannel(abf)
}
