% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot_IVChannelWithIntv}
\alias{Plot_IVChannelWithIntv}
\title{Plot_IVChannelWithIntv plots both current and voltage channels on a same graph along with two dotted lines indicating an interval}
\usage{
Plot_IVChannelWithIntv(abf, intv)
}
\arguments{
\item{abf}{data loaded from abf file}

\item{intv}{a vector of 2 numerics describing an interval}
}
\value{
a ggplot object
}
\description{
Plot_IVChannelWithIntv plots both current and voltage channels on a same graph along with two dotted lines indicating an interval
}
\examples{
intv <- FindSamplingInterval(abf); Plot_IVChannelWithIntv(abf, intv)
}
