% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot_IVSummary}
\alias{Plot_IVSummary}
\title{Plot_IVSummary plots a current-voltage summary}
\usage{
Plot_IVSummary(df_summary, err_bar_width = 1.5)
}
\arguments{
\item{df_summary}{a data.frame can be easily obtained from AllSamples_IVSummary}

\item{err_bar_width}{OPTIONAL, width of error bar}
}
\value{
a ggplot object
}
\description{
Plot_IVSummary plots a current-voltage summary
}
\examples{
See Example2.R
}
