% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotAll_Channel}
\alias{PlotAll_Channel}
\title{PlotAll_Channel plots same channel for all abf data in abf_list}
\usage{
PlotAll_Channel(abf_list, chan_id, label = TRUE, uniform_y = TRUE)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{chan_id}{channel id to plot}

\item{label}{OPTIONAL, determines if labels are added to subplots}

\item{uniform_y}{OPTIONAL, determines if y axes are unified to same scale}
}
\value{
a ggplot object
}
\description{
PlotAll_Channel plots same channel for all abf data in abf_list
}
\examples{
abf_list <- abf2.load_in_folder(path, files); PlotAll_Channel(abf_list, 1)
}
