% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotAll_ChannelWithIntv}
\alias{PlotAll_ChannelWithIntv}
\title{PlotAll_ChannelWithIntv plots same channel for all abf data in abf_list as well as intervals}
\usage{
PlotAll_ChannelWithIntv(abf_list, intv_list, chan_id, label = TRUE,
  title_list = NULL, uniform_y = TRUE)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{intv_list}{a list of intervals}

\item{chan_id}{channel id to plot}

\item{label}{OPTIONAL, determines if labels are added to subplots}

\item{title_list}{}

\item{uniform_y}{OPTIONAL, determines if y axes are unified to same scale}
}
\value{
a ggplot object
}
\description{
PlotAll_ChannelWithIntv plots same channel for all abf data in abf_list as well as intervals
}
\examples{
See PlotAll_Channel and Plot_ChannelWithIntv
}
