% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotAll_IVChannel}
\alias{PlotAll_IVChannel}
\title{PlotAll_IVChannel plots both current and voltage channels for all abf data}
\usage{
PlotAll_IVChannel(abf_list, title_list = NULL, uniform_y = TRUE)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{title_list}{OPTIONAL, a list of titles corresponds to abf_list}

\item{uniform_y}{OPTIONAL, determines if y axes are unified to same scale}
}
\value{
a list of ggplot objects
}
\description{
PlotAll_IVChannel plots both current and voltage channels for all abf data
}
\examples{
see Plot_IVChannel and PlotALL_Channel
}
