% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotAll_IVChannelWithIntv}
\alias{PlotAll_IVChannelWithIntv}
\title{PlotAll_IVChannelWithIntv plots both current and voltage channels for all abf data as well as intervals}
\usage{
PlotAll_IVChannelWithIntv(abf_list, intv_list, title_list = NULL,
  uniform_y = TRUE)
}
\arguments{
\item{abf_list}{a list of abf data}

\item{intv_list}{a list of intervals}

\item{title_list}{OPTIONAL, a list of titles corresponds to abf_list}

\item{uniform_y}{OPTIONAL, determines if y axes are unified to same scale}
}
\value{
a list of ggplot objects
}
\description{
PlotAll_IVChannelWithIntv plots both current and voltage channels for all abf data as well as intervals
}
\examples{
see Plot_IVChannelWithIntv and PlotALL_Channel
}
