% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotAll_IVSummary}
\alias{PlotAll_IVSummary}
\title{PlotAll_IVSummary plots a list of IVSummary}
\usage{
PlotAll_IVSummary(df_summary_list, err_bar_width = 1.5)
}
\arguments{
\item{df_summary_list}{a list of data.frame. IVSummary can be obtained from AllSamples_IVSummary}

\item{err_bar_width}{OPTIONAL, width of error bar}
}
\value{
a ggplot object
}
\description{
PlotAll_IVSummary plots a list of IVSummary
}
\examples{
See Example2.R
}
