% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{RemoveEpisode}
\alias{RemoveEpisode}
\title{RemoveEpisode removes episodes from a channel or a list of channel}
\usage{
RemoveEpisode(ByChannel, episode_idx)
}
\arguments{
\item{ByChannel}{a channel or a list of channels}

\item{episode_idx}{a list or vector of episode indices to be removed}
}
\value{
a channel or a list of channels with the desired episodes removed
}
\description{
RemoveEpisode removes episodes from a channel or a list of channel
}
\examples{
abf$ByChannel <- RemoveEpisode(abf$ByChannel, c(1,3))
}
