% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{select.samples}
\alias{select.samples}
\title{Select samples from data index}
\usage{
select.samples(dataindex, oocyteNo, gene, buffer)
}
\arguments{
\item{dataindex}{: dataindex loaded from csv}

\item{oocyteNo}{: oocyte number}

\item{gene}{: gene name}

\item{buffer}{: buffer name}
}
\value{
Selected samples that satisfy conditions provided
}
\description{
Select samples from data index
}
\examples{
select.samples(loaded_index, c(5:10), "AtNRT1.1", "50mM Na")
}
