% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.best_interval}
\alias{vclamp.best_interval}
\title{Guess the best sampling interval for obtaining I and V data}
\usage{
vclamp.best_interval(sample_channels, interval_pts, current_chan_id,
  voltage_chan_id, target_voltages, voltage_threshold, restriction_div)
}
\arguments{
\item{sample_channels}{a list of channels contain I and V data, abf2data$ByChannel}

\item{interval_pts}{desired interval size}

\item{current_chan_id}{channel id for current, usually 1}

\item{voltage_chan_id}{channel id for voltage, usually 2}

\item{target_voltages}{TEVC protocal voltage settings}

\item{voltage_threshold}{maximum error from protocal voltage settings}

\item{restriction_div}{controls sampling area}
}
\value{
a sampling interval as vector of numeric
}
\description{
Guess the best sampling interval for obtaining I and V data
}
