% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.channel_mean}
\alias{vclamp.channel_mean}
\title{Calculate channel mean}
\usage{
vclamp.channel_mean(sample_list, intervals, chan_id)
}
\arguments{
\item{sample_list}{a list of channels}

\item{chan_id}{}
}
\value{
a data.frame, the episodic averages (as columns) of a perticular channel (given by chan_id) of all samples (as rows)
}
\description{
Calculate channel mean
}
