% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.combine_channel_mean}
\alias{vclamp.combine_channel_mean}
\title{Combine voltage and current mean to a data frame}
\usage{
vclamp.combine_channel_mean(voltage_mean, current_mean)
}
\arguments{
\item{voltage_mean}{from vclamp.channel_mean}

\item{current_mean}{from vclamp.channel_mean}
}
\value{
a data frame with Voltage, SEM Voltage, Current, SEM Current
}
\description{
Combine voltage and current mean to a data frame
}
