% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.individual_channel_mean}
\alias{vclamp.individual_channel_mean}
\title{Combines voltage and current means for each sample}
\usage{
vclamp.individual_channel_mean(voltage_mean, current_mean)
}
\arguments{
\item{voltage_mean}{from vclamp.channel_mean}

\item{current_mean}{from vclamp.channel_man}
}
\value{
a data frame contins voltage mean and current mean for each sample
}
\description{
Combines voltage and current means for each sample
}
