% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.individual_IorV_mean}
\alias{vclamp.individual_IorV_mean}
\title{Combine current or voltage means}
\usage{
vclamp.individual_IorV_mean(IorVmeans)
}
\arguments{
\item{IorVmeans}{a data frame with voltage or current mean}
}
\description{
Combine current or voltage means
}
