% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.plot_channel}
\alias{vclamp.plot_channel}
\title{Plot channel}
\usage{
vclamp.plot_channel(sample_channels, chan_id, xy_info, plottitle)
}
\arguments{
\item{sample_channels}{a list of channels}

\item{plottitle}{}
}
\description{
Plot channel
}
