% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.plot_interval}
\alias{vclamp.plot_interval}
\title{Plot sample channels with interval shown}
\usage{
vclamp.plot_interval(sample_channels, interval, xy_info)
}
\arguments{
\item{sample_channels}{a list of channels}

\item{interval}{a vector of numeric, can be obtained from vclamp.best_interval}

\item{xy_info}{can be obtained from vclamp.plot_xy_info}
}
\description{
Plot sample channels with interval shown
}
