% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.plot_sample_mean}
\alias{vclamp.plot_sample_mean}
\title{Plot sample mean}
\usage{
vclamp.plot_sample_mean(df, current_chan_id = 1, voltage_chan_id = 2)
}
\arguments{
\item{df}{result data frame from vclamp.channel_mean}

\item{voltage_chan_id}{}
}
\description{
Plot sample mean
}
