% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.plot_xy_info}
\alias{vclamp.plot_xy_info}
\title{Get xy_info used for vclamp.plot_channel}
\usage{
vclamp.plot_xy_info(abfdata, time_unit)
}
\arguments{
\item{time_unit}{time unit for x axis, can be "ms", "s" or "ticks"}
}
\description{
Get xy_info used for vclamp.plot_channel
}
