% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.sample_mean}
\alias{vclamp.sample_mean}
\title{Returns a data.frame, the episodic averages (as columns) of channels (as rows) in the given interval of the sample}
\usage{
vclamp.sample_mean(sample_channels, interval)
}
\arguments{
\item{interval}{}
}
\value{
a data frame as described
}
\description{
Returns a data.frame, the episodic averages (as columns) of channels (as rows) in the given interval of the sample
}
