% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_func.R
\name{vclamp.voltage_setting}
\alias{vclamp.voltage_setting}
\title{Get protocol voltage settings from abf data}
\usage{
vclamp.voltage_setting(abfdata)
}
\arguments{
\item{abfdata}{loaded from abf2.load}
}
\value{
protocol voltage settings for each episode as a vector of numeric
}
\description{
Get protocol voltage settings from abf data
}
